+++ Trabajo Practico 1/FuncionesAritmeticas.h
@@ -1,7 +1,49 @@
-int ObtengoNumero();
-int Suma(int numero1,int numero2);
-int Resta(int numero1, int numero2);
-int Multiplicacion(int factor1,int factor2);
+/** \brief Pide al usuario que ingrese un numero y lo guarda en la variable correspondiente.
+ *
+ * \return Numero.
+ *
+ */
+float ObtengoNumero();
+
+/** \brief Suma dos operadores.
+ *
+ * \param Primer operador.
+ * \param Segundo operador.
+ * \return Resultado de la suma de los dos operadores.
+ *
+ */
+float Suma(float numero1,float numero2);
+/** \brief Resta dos numeros.
+ *
+ * \param Primer operador.
+ * \param Segundo operador.
+ * \return Resultado de la resta de los dos operadores.
+ *
+ */
+float Resta(float numero1, float numero2);
+
+/** \brief Multiplica dos coeficientes.
+ *
+ * \param Primer factor.
+ * \param Segundo factor.
+ * \return Producto de la multiplicación de los dos operadores.
+ *
+ */
+double Multiplicacion(float factor1,float factor2);
+/** \brief Divide dos numeros.
+ * \param Error division por 0
+ * \param Dividendo.
+ * \param Divisor.
+ * \return Cociente de la división de los dos operadores.
+ *
+ */
 float Division(float dividendo,float divisor);
+
+/** \brief Calcula el factorial de un entero.
+ *
+ * \param Numero entero.
+ * \return Resultado del factorial de ese numero entero.
+ *
+ */
 long long int Factorial(long long int primerOperando);
